[add delete and flush methods to cache; getDetail methods; add details to createThing; update detail model;](https://github.com/bensbits91/things-server/commit/de87241d3e081f96b0a96fe70b98dbf56ac49069) 

[add compound unique index to thingSchema; return 409 conflict error](https://github.com/bensbits91/things-server/commit/4e51aa0afbbaeef2a6c965179305db1a191471b9)

[custom error classes; global error handler](https://github.com/bensbits91/things-server/commit/58b3977a227f25ff02265bf07374aa50bb74f596)

[add error handling to all services; remove try/catch from all controllers; remove user and registration files](https://github.com/bensbits91/things-server/commit/761af1484fca7babdb63f0aa804f97d001df0ba1) 

[add unique compound index in Detail model; update stauts enum and add statusMap; add DuplicateKeyError; normalize thing details on fetch; return 401 when session ends](https://github.com/bensbits91/things-server/commit/9bdb9ad7f0a0984149fbae1b63ef46ac06874168) 

[add Status; add script to seed statuses](https://github.com/bensbits91/things-server/commit/a255c03505f357742661710057b34afca03ccc85)

+++ src/services/detailService.js
@@ -1,28 +1,81 @@
 const Detail = require('../models/detailModel');
+const { DuplicateKeyError, NotFoundError, ValidationError } = require('../errors/errors');
 
 class DetailService {
    async createDetail(detailData) {
-      console.log('\n\nbb ~ detailData:', detailData, '\n\n');
-      const detail = new Detail(detailData);
-      await detail.save();
-      return detail;
+      if (!detailData) {
+         throw new ValidationError('detailData is required');
+      }
+      try {
+         const detail = new Detail(detailData);
+         await detail.save();
+         return detail;
+      } catch (error) {
+         if (error.code === 11000) {
+            console.error('\n\n\n\nDuplicate entry for type, name, and external_id combination');
+            console.info('bb ~ detailService.js ~ detailData:', detailData, '\n\n\n\n');
+            throw new DuplicateKeyError(
+               'Duplicate entry for type, name, and external_id combination'
+            );
+         }
+         throw error;
+      }
    }
 
    async getDetail(id) {
-      return Detail.findById(id);
+      if (!id) {
+         throw new ValidationError('ID is required');
+      }
+      try {
+         return Detail.findById(id);
+      } catch (error) {
+         if (error.code === 404) {
+            throw new NotFoundError('Detail not found');
+         }
+         throw error;
+      }
    }
 
-   // async getDetailsByUser(userId) {
-   //    console.log('\n\nbb ~ detailService.js ~ userId:', userId, '\n\n');
-   //    return Detail.find({ userId: userId });
-   // }
+   async getDetailByExternalId(externalId) {
+      if (!externalId) {
+         throw new ValidationError('externalId is required');
+      }
+      try {
+         return Detail.findOne({ external_id: externalId });
+      } catch (error) {
+         if (error.code === 404) {
+            throw new NotFoundError('Detail not found');
+         }
+         throw error;
+      }
+   }
 
    async updateDetail(id, updateData) {
-      return Detail.findByIdAndUpdate(id, updateData, { new: true });
+      if (!query || !type || !results) {
+         throw new ValidationError('Missing required parameters');
+      }
+      try {
+         return Detail.findByIdAndUpdate(id, updateData, { new: true });
+      } catch (error) {
+         if (error.code === 404) {
+            throw new NotFoundError('Detail not found for update');
+         }
+         throw error;
+      }
    }
 
    async deleteDetail(id) {
-      return Detail.findByIdAndDelete(id);
+      if (!id) {
+         throw new ValidationError('ID is required');
+      }
+      try {
+         return Detail.findByIdAndDelete(id);
+      } catch (error) {
+         if (error.code === 404) {
+            throw new NotFoundError('Detail not found for deletion');
+         }
+         throw error;
+      }
    }
 }
